Added Utility file for dlt. and move related files from resources.py into new utils
